requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
python-dotenv
requests
pydantic
chromadb 